fix(ui): resolve lint errors and clean up screenshots
- Move needsDetailSheet to widget-utils.ts (react-refresh rule)
- Fix setState-in-effect in WidgetGrid and QrLoginPage
- Move familyTypes inside useMemo callback
- Ignore dev-dist in eslint config
- Remove unused effectiveDetailId variable
