+++ ui/src/components/dashboard/widget-utils.ts
@@ -1,0 +1,10 @@
+/** Does this equipment type need a detail bottom sheet on mobile? */
+export function needsDetailSheet(equipmentType: string): boolean {
+  return [
+    "light_dimmable",
+    "light_color",
+    "shutter",
+    "thermostat",
+    "heater",
+  ].includes(equipmentType);
+}
