+++ ui/src/pages/QrLoginPage.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useState } from "react";
+import { useEffect, useMemo, useState } from "react";
 import { useSearchParams, useNavigate, Link } from "react-router-dom";
 import { useTranslation } from "react-i18next";
 import { Loader2, AlertCircle } from "lucide-react";
@@ -14,12 +14,10 @@ export function QrLoginPage() {
   const [error, setError] = useState<string | null>(null);
 
   const token = searchParams.get("token");
+  const invalidMessage = useMemo(() => (!token ? t("qrLogin.invalidToken") : null), [token, t]);
 
   useEffect(() => {
-    if (!token) {
-      setError(t("qrLogin.invalidToken"));
-      return;
-    }
+    if (!token) return;
 
     let cancelled = false;
 
@@ -50,17 +48,19 @@ export function QrLoginPage() {
     }
   }, [isAuthenticated, error, navigate]);
 
+  const displayError = invalidMessage ?? error;
+
   return (
     <div className="flex items-center justify-center min-h-screen bg-background">
       <div className="flex flex-col items-center gap-6 p-8">
         <SowelLogo size={48} />
         <h1 className="text-[20px] font-semibold text-text">Sowel</h1>
 
-        {error ? (
+        {displayError ? (
           <div className="flex flex-col items-center gap-4">
             <div className="flex items-center gap-2 text-error">
               <AlertCircle size={18} strokeWidth={1.5} />
-              <span className="text-[14px]">{error}</span>
+              <span className="text-[14px]">{displayError}</span>
             </div>
             <Link
               to="/login"
